# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: NXG Ecommerce API
  description: 'This is a NXG Ecommerce module. You can find out more about NXG Ecommerce at [http://nxgecommerce.in](http://nxgecommerce.io) or on [www.nxgecommerce.in, #nxgecommerce].'
  termsOfService: http://nxgecommerce/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: https://localhost:8080/nxgecommerce/api/v1
  - url: http://localhost:8080/nxgecommerce/api/v1
...
paths:
...
  /product:
    post:
      tags:
        - Product Information
      summary: Adding a new product
      requestBody:
        description: product details that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        '201':
          description: Successful
          content: {}      
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}          
        '502':
          description: Bad Gatway
          content: {}               
      x-codegen-request-body-name: body
    get:
      tags:
        - Product Information
      summary:  Getting all products details
      responses:
        '200':
          description:  A list of all products details
          content:
            applicantion/json:
              schema:
                $ref: '#/components/schemas/Product'
        '201':
          description: Successful
          content: {}
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}
        '502':
          description: Bad Gatway
      x-codegen-request-body-name: body
  /product/{productId}:
    get:
      tags:
        - Product Information
      summary: Getting product details by id
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: A list of a product details
          content:
            applicantion/json:
              schema:
                $ref: '#/components/schemas/Product'
        '201':
          description: Successful
          content: {}
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}
        '502':
          description: Bad Gatway
      x-codegen-request-body-name: body   
    put:
        tags:
          - Product Information
        summary: Editing product details
        parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
        requestBody:
          description: Product details that need to be added
          content:
            applicantion/json:
              schema:
                $ref: '#/components/schemas/Product'
          required: true      
        responses:
          '200':
            description: A list of a product details
            content:
              applicantion/json:
                schema:
                  $ref: '#/components/schemas/Product'
          '201':
            description: Successful
            content: {}
          '400':
            description: Bad request input
            content: {}
          '404':
            description: Data not found
            content: {}
          '500':
            description: Invalid input
            content: {}
          '502':
            description: Bad Gatway
        x-codegen-request-body-name: body
    delete:
      tags:
        - Product Information
      summary: Delete a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Successful
          content: {}      
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}          
        '502':
          description: Bad Gatway
          content: {}               
      x-codegen-request-body-name: body    
components:
  schemas:
...
    Product:
      type: object
      required:
        - productId
        - categoryId
        - title
        - price
        - cPrice
        - taxPercent
        - qty
        - sizeType
        - description
        - skuCode
        - barcode
        - isVariants
        - productUrl
        - searchTerms
        - seoTitle
        - meta
        - keyword
      properties:
        productId:
          type: integer
          maxLength: 11
        categoryId:
          type: integer
          maxLength: 11
        title:
          type: string
          maxLength: 200
        price:
          type: number
          format: float
        cPrice:
          type: number
          format: float
        taxPercent:
          type: number
          format: float
        qty:
          type: integer
          maxLength: 10
        sizeType:
          type: integer
          maxLength: 1
        description:
          type: string
        skuCode:
          type: string
        barcode:
          type: string
        isVariants:
          type: integer
          maxLength: 11
        productUrl:
          type: string
        searchTerms:
          type: string
        seoTitle:
          type: string
          maxLength: 250
        meta:
          type: string
          maxLength: 170
        keyword:
          type: string
          maxLength: 200
        productImage:
          type: array
          items:
            "$ref": "#/components/schemas/ProductImage"
        productVariant:
          type: array
          items:
            "$ref": "#/components/schemas/ProductVariant"
    ProductImage:
      type: object
      properties:
        id:
          type: integer
          maxLength: 11
        productId:
          type: integer
          maxLength: 11
        imageName:
          type: string
          maxLength: 200
        url:
          type: string
        isPrimary:
          type: integer
          maxLength: 1
    ProductVariant:
      type: object
      properties:
        productVariantId:
          type: integer
          maxLength: 11
        productId:
          type: integer
          maxLength: 11
        variantOptionValueId1:
          type: integer
          maxLength: 11
        variantOptionValueId2:
          type: integer
          maxLength: 11
        variantCombinationName:
          type: string
          maxLength: 120
        price:
          type: number
          format: float
        sku:
          type: string
          maxLength: 60
        quantity:
          type: integer
          maxLength: 11
        barcode:
          type: string
          maxLength: 120
        isDefault:
          type: boolean
        userId:
          type: integer
          maxLength: 11
...